fix caas vitest
