+++ lib/metrics/index.js
@@ -6,7 +6,14 @@ const { getStringFromEnv } = require('@opentelemetry/core')
 const { resourceFromAttributes } = require('@opentelemetry/resources')
 const { AggregationTemporality, MeterProvider, PeriodicExportingMetricReader } = require('@opentelemetry/sdk-metrics')
 
-const { getDynatraceMetadata, getCredsForDTAsUPS, getCredsForCLSAsUPS, augmentCLCreds, createCaaSExporter, _require } = require('../utils')
+const {
+  getDynatraceMetadata,
+  getCredsForDTAsUPS,
+  getCredsForCLSAsUPS,
+  augmentCLCreds,
+  createCaaSExporter,
+  _require
+} = require('../utils')
 
 const _protocol2module = {
   grpc: '@opentelemetry/exporter-metrics-otlp-grpc',
